indexTopicHandler: index task title even if it's not known as 'loaded'
The 'task' field, containing the title, is a Modell Aachen extension,
hence there's no default type for it. We define the type in _EXTRAS
settings, but if, for some reason, the ATP is inited in the
indexTopicHandler without an appropriate _EXTRAS setting, the task field
will not be loaded normally for the remainder of the index run.

To fix, make sure to fall back to the unloaded fields when reading the
task title.

Ticket: none