Make AJAX updating configurable
By default, updating actions happens using a normal form submit. Set
ACTIONTRACKERPLUGIN_UPDATEAJAX to 1 to make it use AJAX updating
instead. Please be aware that this mode doesn't immediately update the
action view on the current page; a reload is required for that.

This version introduces a new ACTIONEDIT macro (params web,topic,uid) to
conveniently generate an edit button for actions that respects the value
of the UPDATEAJAX preference.

Ticket: #3902